WIP - Replace Travis and Appveyor with GitHub Actions
This replaces Travis (for Linux & Mac) and Appveyor (for Windows) with
GitHub Actions.

This also expands the automation from just running tests and attaching
release binaries to:

 - Creating a draft release on GitHub
 - Extracting release notes from CHANGELOG.md, attaching them to the
   draft release
 - Building and testing all combinations of environments, release
   types, and library types
 - Attaching release assets
 - Fully publishing the draft release on GitHub
 - Publishing to NPM
 - Publishing to Docker Hub

All of these steps will be run when a new release tag is created.  If
any step fails (tests, for example, or building a certain release
asset), then the workflow will be aborted.

Change-Id: Ic53eef60a8587c5d1487769a0cefaa16eb9b46e7