Roll src/packager/tools/clang/ 723b25997..aee8160f9 (804 commits)
https://chromium.googlesource.com/chromium/src/tools/clang/+log/723b25997f0a..aee8160f9ac4

An upgrade to clang is necessary to get the build to work in MSVS
2019.  The older clang download tools don't recognize the MSVS 2019
environment.  (Even though we aren't using clang for a Windows build,
the clang tools still run as part of gclient sync.)

This commit upgrades clang, which is working on Linux and Mac.
Windows builds with MSVS 2019 are not quite working yet.

Issue #867

Created with:
  roll-dep src/packager/tools/clang

Change-Id: Ic18e4c598bf010de38ab04f8ad8bcbbb0ab77ccd