Fix build issues with new clang
An upgrade to clang is necessary to get the build to work in MSVS
2019.  The older clang download tools don't recognize the MSVS 2019
environment.  (Even though we aren't using clang for a Windows build,
the clang tools still run as part of gclient sync.)

This commit fixes build issues that arise with the newer version of
clang before we actually upgrade clang in a subsequent commit.
The new clang has several new warnings that are treated as errors, and
this commit disables those warnings.  Most of the warnings appear in
code we don't control anyway.

Issue #867

Change-Id: Ib4c823b6d22e107ea45c505565bed1c944e1f933