Fix link issues on macOS
Newer clang releases fail when a linker path does not exist.  To
enable a newer clang release to be used, we need to add the expected
path packager/third_party/libc++-static.

We are using very old versions of the Chromium build system, and
Chromium build from that era wants a libc++-static library for Mac.
This has since been removed from Chromium, because it is not necessary
in newer macOS SDKs.  Even though we target a new enough version of
macOS not to need the libc++-static hack, the build system still tells
the compiler to expect this folder to exist.  The contents are not
actually needed, though.

Change-Id: I1192299b281a0208c3b41a26feee94175c3701b0