ci: Parallelize and cache builds
Previously, we disabled parallel builds in CI because we were running
out of memory.  This re-enables parallel builds with limited
parallelism in linking (since linking debug builds was the major
cause of OOM conditions).

This also enables caching with ccache to speed up the build steps.
This should make CI feedback much faster when small changes are made
to a PR.